+++ packages/core/src/types/IImage.ts
@@ -10,6 +10,8 @@ import { ImageQualityStatus } from '../enums';
 interface IImage {
   /** Image Id */
   imageId: string;
+  /** referenced imageId if this image is derived from */
+  referencedImageId?: string;
   sharedCacheKey?: string;
   /** Whether the image is Pre-scaled during loading */
   isPreScaled?: boolean;