+++ packages/core/src/types/PixelDataTypedArray.ts
@@ -5,3 +5,11 @@ export type PixelDataTypedArray =
   | Uint8Array
   | Int8Array
   | Uint8ClampedArray;
+
+export type PixelDataTypedArrayString =
+  | 'Float32Array'
+  | 'Int16Array'
+  | 'Uint16Array'
+  | 'Uint8Array'
+  | 'Int8Array'
+  | 'Uint8ClampedArray';