+++ packages/tools/src/tools/segmentation/SphereScissorsTool.ts
@@ -1,4 +1,4 @@
-import { cache, getEnabledElement, StackViewport } from '@cornerstonejs/core';
+import { cache, getEnabledElement } from '@cornerstonejs/core';
 import type { Types } from '@cornerstonejs/core';
 
 import { BaseTool } from '../base';
@@ -10,7 +10,7 @@ import {
 } from '../../types';
 
 import { fillInsideSphere } from './strategies/fillSphere';
-import { Events } from '../../enums';
+import { Events, SegmentationRepresentations } from '../../enums';
 import { drawCircle as drawCircleSvg } from '../../drawingSvg';
 import {
   resetElementCursor,
@@ -26,8 +26,12 @@ import {
 } from '../../stateManagement/segmentation';
 
 import { getSegmentation } from '../../stateManagement/segmentation/segmentationState';
-import { LabelmapSegmentationData } from '../../types/LabelmapTypes';
-
+import {
+  LabelmapSegmentationData,
+  LabelmapSegmentationDataVolume,
+  LabelmapSegmentationDataStack,
+} from '../../types/LabelmapTypes';
+import { isVolumeSegmentation } from './strategies/utils/stackVolumeCheck';
 /**
  * Tool for manipulating segmentation data by drawing a sphere in 3d space. It acts on the
  * active Segmentation on the viewport (enabled element) and requires an active
@@ -40,10 +44,14 @@ class SphereScissorsTool extends BaseTool {
   static toolName;
   editData: {
     annotation: any;
-    segmentation: any;
     segmentIndex: number;
     segmentsLocked: number[];
-    segmentationId: string;
+    segmentationRepresentationUID: string;
+    //
+    volumeId: string;
+    referencedVolumeId: string;
+    imageIdReferenceMap: Map<string, string>;
+    //
     toolGroupId: string;
     segmentColor: [number, number, number, number];
     viewportIdsToRender: string[];
@@ -103,7 +111,7 @@ class SphereScissorsTool extends BaseTool {
       );
     }
 
-    const { segmentationRepresentationUID, segmentationId, type } =
+    const { segmentationRepresentationUID, segmentationId } =
       activeSegmentationRepresentation;
     const segmentIndex =
       segmentIndexController.getActiveSegmentIndex(segmentationId);
@@ -114,12 +122,6 @@ class SphereScissorsTool extends BaseTool {
       segmentationRepresentationUID,
       segmentIndex
     );
-
-    const { representationData } = getSegmentation(segmentationId);
-
-    // Todo: are we going to support contour editing with rectangle scissors?
-    const { volumeId } = representationData[type] as LabelmapSegmentationData;
-    const segmentation = cache.getVolume(volumeId);
 
     this.isDrawing = true;
 
@@ -148,19 +150,41 @@ class SphereScissorsTool extends BaseTool {
 
     this.editData = {
       annotation,
-      segmentation,
       centerCanvas: canvasPos,
+      segmentationRepresentationUID,
       segmentIndex,
       segmentsLocked,
       segmentColor,
-      segmentationId,
       toolGroupId,
       viewportIdsToRender,
       handleIndex: 3,
       movingTextBox: false,
       newAnnotation: true,
       hasMoved: false,
-    };
+    } as any;
+
+    const { representationData } = getSegmentation(segmentationId);
+    const labelmapData =
+      representationData[SegmentationRepresentations.Labelmap];
+
+    if (isVolumeSegmentation(labelmapData as LabelmapSegmentationData)) {
+      const { volumeId } = labelmapData as LabelmapSegmentationDataVolume;
+      const segmentation = cache.getVolume(volumeId);
+
+      this.editData = {
+        ...this.editData,
+        volumeId,
+        referencedVolumeId: segmentation.referencedVolumeId,
+      };
+    } else {
+      const { imageIdReferenceMap } =
+        labelmapData as LabelmapSegmentationDataStack;
+
+      this.editData = {
+        ...this.editData,
+        imageIdReferenceMap,
+      };
+    }
 
     this._activateDraw(element);
 
@@ -227,10 +251,9 @@ class SphereScissorsTool extends BaseTool {
       annotation,
       newAnnotation,
       hasMoved,
-      segmentation,
       segmentIndex,
+      segmentationRepresentationUID,
       segmentsLocked,
-      segmentationId,
     } = this.editData;
     const { data } = annotation;
     const { viewPlaneNormal, viewUp } = annotation.metadata;
@@ -246,24 +269,19 @@ class SphereScissorsTool extends BaseTool {
     resetElementCursor(element);
 
     const enabledElement = getEnabledElement(element);
-    const { viewport } = enabledElement;
-
-    this.editData = null;
-    this.isDrawing = false;
-
-    if (viewport instanceof StackViewport) {
-      throw new Error('Not implemented yet');
-    }
 
     const operationData = {
+      ...this.editData,
       points: data.handles.points,
-      volume: segmentation,
       segmentIndex,
+      segmentationRepresentationUID,
       segmentsLocked,
-      segmentationId,
       viewPlaneNormal,
       viewUp,
     };
+
+    this.editData = null;
+    this.isDrawing = false;
 
     this.applyActiveStrategy(enabledElement, operationData);
   };