+++ packages/tools/src/tools/segmentation/strategies/fillSphere.ts
@@ -1,21 +1,15 @@
 import type { Types } from '@cornerstonejs/core';
-import { utilities as csUtils } from '@cornerstonejs/core';
+import { cache, utilities as csUtils } from '@cornerstonejs/core';
 
 import { triggerSegmentationDataModified } from '../../../stateManagement/segmentation/triggerSegmentationEvents';
 import { pointInSurroundingSphereCallback } from '../../../utilities';
 import isWithinThreshold from './utils/isWithinThreshold';
+import { LabelmapToolOperationData } from '../../../types';
+import { getStrategyData } from './utils/getStrategyData';
+import { isVolumeSegmentation } from './utils/stackVolumeCheck';
 
-type OperationData = {
+type OperationData = LabelmapToolOperationData & {
   points: [Types.Point3, Types.Point3, Types.Point3, Types.Point3];
-  volume: Types.IImageVolume;
-  imageVolume: Types.IImageVolume;
-  segmentIndex: number;
-  segmentationId: string;
-  segmentsLocked: number[];
-  viewPlaneNormal: Types.Point3;
-  viewUp: Types.Point3;
-  strategySpecificConfiguration: any;
-  constraintFn: () => boolean;
 };
 
 function fillSphere(
@@ -26,31 +20,36 @@ function fillSphere(
 ): void {
   const { viewport } = enabledElement;
   const {
-    volume: segmentation,
     segmentsLocked,
     segmentIndex,
-    imageVolume,
     strategySpecificConfiguration,
-    segmentationId,
     points,
   } = operationData;
 
-  const { imageData, dimensions } = segmentation;
-  const scalarData = segmentation.getScalarData();
+  const strategyData = getStrategyData({ operationData, viewport });
+
+  if (!strategyData) {
+    console.warn('No data found for fillSphere');
+    return;
+  }
+
+  const { imageScalarData, segmentationImageData, segmentationScalarData } =
+    strategyData;
+
   const scalarIndex = [];
 
   let callback;
 
   if (threshold) {
-    callback = ({ value, index, pointIJK }) => {
+    callback = ({ value, index }) => {
       if (segmentsLocked.includes(value)) {
         return;
       }
 
       if (
-        isWithinThreshold(index, imageVolume, strategySpecificConfiguration)
+        isWithinThreshold(index, imageScalarData, strategySpecificConfiguration)
       ) {
-        scalarData[index] = segmentIndex;
+        segmentationScalarData[index] = segmentIndex;
         scalarIndex.push(index);
       }
     };
@@ -59,30 +58,36 @@ function fillSphere(
       if (segmentsLocked.includes(value)) {
         return;
       }
-      scalarData[index] = segmentIndex;
+      segmentationScalarData[index] = segmentIndex;
       scalarIndex.push(index);
     };
   }
 
   pointInSurroundingSphereCallback(
-    imageData,
+    segmentationImageData,
     [points[0], points[1]],
     callback,
     viewport as Types.IVolumeViewport
   );
 
-  // Since the scalar indexes start from the top left corner of the cube, the first
-  // slice that needs to be rendered can be calculated from the first mask coordinate
-  // divided by the zMultiple, as well as the last slice for the last coordinate
-  const zMultiple = dimensions[0] * dimensions[1];
-  const minSlice = Math.floor(scalarIndex[0] / zMultiple);
-  const maxSlice = Math.floor(scalarIndex[scalarIndex.length - 1] / zMultiple);
-  const sliceArray = Array.from(
-    { length: maxSlice - minSlice + 1 },
-    (v, k) => k + minSlice
-  );
+  const dimensions = segmentationImageData.getDimensions();
 
-  triggerSegmentationDataModified(segmentationId, sliceArray);
+  let sliceArray;
+  if (isVolumeSegmentation(operationData)) {
+    // Since the scalar indexes start from the top left corner of the cube, the first
+    // slice that needs to be rendered can be calculated from the first mask coordinate
+    // divided by the zMultiple, as well as the last slice for the last coordinate
+    const zMultiple = dimensions[0] * dimensions[1];
+    const minSlice = Math.floor(scalarIndex[0] / zMultiple);
+    const maxSlice = Math.floor(
+      scalarIndex[scalarIndex.length - 1] / zMultiple
+    );
+    sliceArray = Array.from(
+      { length: maxSlice - minSlice + 1 },
+      (v, k) => k + minSlice
+    );
+    triggerSegmentationDataModified(operationData.volumeId, sliceArray);
+  }
 }
 
 /**
@@ -108,15 +113,20 @@ export function thresholdInsideSphere(
   enabledElement: Types.IEnabledElement,
   operationData: OperationData
 ): void {
-  const { volume, imageVolume } = operationData;
+  if (isVolumeSegmentation(operationData)) {
+    const { referencedVolumeId, volumeId } = operationData;
 
-  if (
-    !csUtils.isEqual(volume.dimensions, imageVolume.dimensions) ||
-    !csUtils.isEqual(volume.direction, imageVolume.direction)
-  ) {
-    throw new Error(
-      'Only source data the same dimensions/size/orientation as the segmentation currently supported.'
-    );
+    const imageVolume = cache.getVolume(referencedVolumeId);
+    const segmentation = cache.getVolume(volumeId);
+
+    if (
+      !csUtils.isEqual(segmentation.dimensions, imageVolume.dimensions) ||
+      !csUtils.isEqual(segmentation.direction, imageVolume.direction)
+    ) {
+      throw new Error(
+        'Only source data the same dimensions/size/orientation as the segmentation currently supported.'
+      );
+    }
   }
 
   fillSphere(enabledElement, operationData, true, true);