+++ packages/tools/src/types/ScrollOptions.ts
@@ -4,6 +4,7 @@ type ScrollOptions = {
   volumeId?: string;
   debounceLoading?: boolean;
   loop?: boolean;
+  scrollSlabs?: boolean;
 };
 
 export default ScrollOptions;