+++ packages/tools/src/types/index.ts
@@ -58,6 +58,22 @@ import type { ContourSegmentationData } from './ContourTypes';
 import type IAnnotationManager from './IAnnotationManager';
 import type AnnotationGroupSelector from './AnnotationGroupSelector';
 import type { Statistics } from './CalculatorTypes';
+import {
+  LabelmapToolOperationData,
+  LabelmapToolOperationDataStack,
+  LabelmapToolOperationDataVolume,
+} from './LabelmapToolOperationData';
+
+// Splines
+import type { CardinalSplineProps } from './CardinalSplineProps';
+import type { ClosestControlPoint } from './ClosestControlPoint';
+import type { ClosestPoint } from './ClosestPoint';
+import type { ClosestSplinePoint } from './ClosestSplinePoint';
+import type { ControlPointInfo } from './ControlPointInfo';
+import type { ISpline } from './ISpline';
+import type { SplineCurveSegment } from './SplineCurveSegment';
+import type { SplineLineSegment } from './SplineLineSegment';
+import type { SplineProps } from './SplineProps';
 
 export type {
   // AnnotationState
@@ -124,4 +140,17 @@ export type {
   ContourSegmentationData,
   //Statistics
   Statistics,
+  LabelmapToolOperationData,
+  LabelmapToolOperationDataStack,
+  LabelmapToolOperationDataVolume,
+  // Splines
+  CardinalSplineProps,
+  ClosestControlPoint,
+  ClosestPoint,
+  ClosestSplinePoint,
+  ControlPointInfo,
+  ISpline,
+  SplineCurveSegment,
+  SplineLineSegment,
+  SplineProps,
 };