+++ packages/tools/src/utilities/math/point/index.ts
@@ -1,3 +1,3 @@
-import distanceToPoint from './distanceToPoint';
-
-export { distanceToPoint };
+export { default as distanceToPoint } from './distanceToPoint';
+export { default as distanceToPointSquared } from './distanceToPointSquared';
+export { default as mirror } from './mirror';